refactor: LadderGameResults 패키지 이동 및 Record 로 변환
