+++ src/main/java/dto/LadderGameResult.java
@@ -1,4 +1,4 @@
-package domain;
+package dto;
 
 public record LadderGameResult(String playerName, String giftName) {
 }
